+++ sdk/servicebus/azure-messaging-servicebus/src/main/java/com/azure/messaging/servicebus/administration/ServiceBusAdministrationClient.java
@@ -39,6 +39,7 @@ import com.azure.messaging.servicebus.administration.implementation.models.Subsc
 import com.azure.messaging.servicebus.administration.implementation.models.SubscriptionDescriptionFeed;
 import com.azure.messaging.servicebus.administration.implementation.models.TopicDescriptionEntry;
 import com.azure.messaging.servicebus.administration.implementation.models.TopicDescriptionFeed;
+import com.azure.messaging.servicebus.administration.implementation.models.RuleDescription;
 import com.azure.messaging.servicebus.administration.models.CreateQueueOptions;
 import com.azure.messaging.servicebus.administration.models.CreateRuleOptions;
 import com.azure.messaging.servicebus.administration.models.CreateSubscriptionOptions;
@@ -255,7 +256,7 @@ public final class ServiceBusAdministrationClient {
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
     public RuleProperties createRule(String topicName, String subscriptionName, String ruleName) {
-        return createRule(topicName, subscriptionName, ruleName, null);
+        return createRule(topicName, subscriptionName, ruleName, new CreateRuleOptions());
     }
 
     /**
@@ -330,11 +331,11 @@ public final class ServiceBusAdministrationClient {
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
     public SubscriptionProperties createSubscription(String topicName, String subscriptionName) {
-        return createSubscription(topicName, subscriptionName, null);
+        return createSubscription(topicName, subscriptionName, new CreateSubscriptionOptions());
     }
 
     /**
-     * Creates a subscription with the {@link SubscriptionProperties}.
+     * Creates a subscription with the {@link CreateSubscriptionOptions}.
      *
      * @param topicName Name of the topic associated with subscription.
      * @param subscriptionName Name of the subscription.
@@ -356,7 +357,7 @@ public final class ServiceBusAdministrationClient {
     }
 
     /**
-     * Creates a queue and returns the created queue in addition to the HTTP response.
+     * Creates a subscription and returns the created subscription in addition to the HTTP response.
      *
      * @param topicName Name of the topic associated with subscription.
      * @param subscriptionName Name of the subscription.
@@ -398,6 +399,73 @@ public final class ServiceBusAdministrationClient {
         return deserializeSubscription(topicName,
             managementClient.getSubscriptions().putSyncWithResponse(topicName, subscriptionName, createEntity,
                 null, contextWithHeaders));
+
+    }
+
+    /**
+     * Creates a subscription with default rule using the {@link CreateSubscriptionOptions} .
+     *
+     * @param topicName Name of the topic associated with subscription.
+     * @param subscriptionName Name of the subscription.
+     * @param ruleName Name of the default rule the subscription should be created with.
+     * @param subscriptionOptions Information about the subscription to create.
+     * @param ruleOptions A {@link CreateRuleOptions} object describing the default rule.
+     *                    If null, then pass-through filter will be created.
+     *
+     * @return Information about the created subscription.
+     * @throws ClientAuthenticationException if the client's credentials do not have access to modify the
+     *     namespace.
+     * @throws HttpResponseException If the request body was invalid, the quota is exceeded, or an error occurred
+     * processing the request.
+     * @throws IllegalArgumentException if {@code topicName} or {@code subscriptionName} are null or empty strings.
+     * @throws ResourceExistsException if a subscription exists with the same topic and subscription name.
+     * @see <a href="https://docs.microsoft.com/rest/api/servicebus/update-entity">Create or Update Entity</a>
+     */
+    @ServiceMethod(returns = ReturnType.SINGLE)
+    public SubscriptionProperties createSubscription(String topicName, String subscriptionName, String ruleName,
+                                                     CreateSubscriptionOptions subscriptionOptions,
+                                                     CreateRuleOptions ruleOptions) {
+        return createSubscriptionWithResponse(topicName, subscriptionName, ruleName, subscriptionOptions,
+            ruleOptions, null).getValue();
+    }
+
+    /**
+     * Creates a subscription with default rule configured and returns the created subscription
+     * in addition to the HTTP response.
+     *
+     * @param topicName Name of the topic associated with subscription.
+     * @param subscriptionName Name of the subscription.
+     * @param ruleName Name of the default rule the subscription should be created with.
+     * @param subscriptionOptions Information about the subscription to create.
+     * @param ruleOptions A {@link CreateRuleOptions} object describing the default rule.
+     *                    If null, then pass-through filter will be created.
+     * @param context Additional context that is passed through the HTTP pipeline during the service call.
+     *
+     * @return The created subscription in addition to the HTTP response.
+     * @throws ClientAuthenticationException if the client's credentials do not have access to modify the
+     *     namespace.
+     * @throws HttpResponseException If the request body was invalid, the quota is exceeded, or an error occurred
+     * processing the request.
+     * @throws NullPointerException if {@code subscriptionOptions} is null.
+     * @throws ResourceExistsException if a subscription exists with the same topic and subscription name.
+     * @see <a href="https://docs.microsoft.com/rest/api/servicebus/update-entity">Create or Update Entity</a>
+     */
+    @ServiceMethod(returns = ReturnType.SINGLE)
+    public Response<SubscriptionProperties> createSubscriptionWithResponse(String topicName, String subscriptionName,
+                                                                           String ruleName,
+                                                                           CreateSubscriptionOptions subscriptionOptions,
+                                                                           CreateRuleOptions ruleOptions,
+                                                                           Context context) {
+        if (ruleOptions == null) {
+            throw LOGGER.logExceptionAsError(new NullPointerException("'CreateRuleOptions' cannot be null."));
+        }
+        Objects.requireNonNull(ruleOptions.getFilter(), "'RuleFilter' cannot be null.");
+        final RuleDescription rule = new RuleDescription()
+            .setAction(ruleOptions.getAction() != null ? EntityHelper.toImplementation(ruleOptions.getAction()) : null)
+            .setFilter(EntityHelper.toImplementation(ruleOptions.getFilter()))
+            .setName(ruleName);
+        subscriptionOptions.setDefaultRule(EntityHelper.toModel(rule));
+        return createSubscriptionWithResponse(topicName, subscriptionName, subscriptionOptions, context);
 
     }